# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Niche Trend and Tag",
    "description": "Trending keywords, tags, quotes and color palettes for Etsy sellers, POD designers, and content creators.",
    "version": "1.0.0",
    "contact": {
      "name": "Tools That Work",
      "url": "https://toolsthatwork.github.io"
    }
  },
  "servers": [
    {
      "url": "https://niche-trend-api.onrender.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API Info",
        "description": "Returns API name, version, and available endpoints.",
        "operationId": "getInfo",
        "responses": {
          "200": {
            "description": "API info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Returns the health status of the API.",
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/niches": {
      "get": {
        "summary": "List All Niches",
        "description": "Returns all 10 supported niche categories with their seed keywords.",
        "operationId": "getNiches",
        "responses": {
          "200": {
            "description": "List of niches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/trends": {
      "get": {
        "summary": "Get Trending Keywords",
        "description": "Returns trending keywords for a given niche. Data refreshes every 6 hours. Supports: productivity, wellness, motivation, fitness, gaming, pets, finance, nature, reading, parenting.",
        "operationId": "getTrends",
        "parameters": [
          {
            "name": "niche",
            "in": "query",
            "required": false,
            "description": "Niche category (e.g. productivity, wellness, motivation)",
            "schema": {
              "type": "string",
              "default": "productivity",
              "enum": ["productivity", "wellness", "motivation", "fitness", "gaming", "pets", "finance", "nature", "reading", "parenting"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trending keywords for the niche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "niche": { "type": "string" },
                    "keywords": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "cached": { "type": "boolean" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tags": {
      "get": {
        "summary": "Get Niche Tags",
        "description": "Returns optimized hashtags and listing tags for a given niche. Perfect for Etsy listings, Instagram, TikTok, and Pinterest.",
        "operationId": "getTags",
        "parameters": [
          {
            "name": "niche",
            "in": "query",
            "required": false,
            "description": "Niche category",
            "schema": {
              "type": "string",
              "default": "productivity",
              "enum": ["productivity", "wellness", "motivation", "fitness", "gaming", "pets", "finance", "nature", "reading", "parenting"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of tags to return (max 50)",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags for the niche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "niche": { "type": "string" },
                    "tags": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quotes": {
      "get": {
        "summary": "Get Niche Quotes",
        "description": "Returns curated quotes for a given niche. Great for social media posts, content creation, and print-on-demand designs.",
        "operationId": "getQuotes",
        "parameters": [
          {
            "name": "niche",
            "in": "query",
            "required": false,
            "description": "Niche category",
            "schema": {
              "type": "string",
              "default": "motivation"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of quotes to return",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quotes for the niche",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "niche": { "type": "string" },
                    "quotes": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/design-quotes": {
      "get": {
        "summary": "Get Design Quotes with Color Palettes",
        "description": "Returns quotes bundled with matching color palettes. Perfect for print-on-demand designers who need ready-to-use design content.",
        "operationId": "getDesignQuotes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of design quotes to return",
            "schema": {
              "type": "integer",
              "default": 5,
              "minimum": 1,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Design quotes with color palettes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "designs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "quote": { "type": "string" },
                          "palette": {
                            "type": "object",
                            "properties": {
                              "name": { "type": "string" },
                              "background": { "type": "string" },
                              "text": { "type": "string" },
                              "accent": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}